# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI документация для фулфилмент сервиса
  version: v1
servers:
- url: http://localhost:8080
  description: Url сервера
...
paths:
  /api/v1/product:
    get:
      tags:
      - Product API
      summary: Поиск товаров
      description: Возвращает список товаров пользователя с фильтром и пагинацией
      operationId: getAllProducts
      parameters:
      - name: search
        in: query
        description: "Строка для поиска товаров по началу наименования, артикула и\
          ли баркода"
        required: false
        schema:
          type: string
        example: кюлоты
      - name: page
        in: query
        description: Номер страницы (начинается с 0)
        required: false
        schema:
          type: integer
          minimum: 0
        example: 0
      - name: limit
        in: query
        description: Количество элементов на странице
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
        example: 5
      responses:
        "200":
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllProductsResponse"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T19:34:23.56Z
                status: 400
                error: Bad request
                message: The limit cannot exceed 30
                path: /api/v1/product
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/product
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't get products
                path: /api/v1/product
      security:
      - TelegramAuth: []
    put:
      tags:
      - Product API
      summary: Обновление информации о товарах
      description: "Обновляет информацию о товарах пользователя в базе данных, дел\
        ая запрос к api Wildberries"
      operationId: updateProductInformation
      responses:
        "200":
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"
        "400":
          description: Недействительный api ключ Wildberries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:04:22.00Z
                status: 400
                error: "Wildberries API Error: unauthorized"
                message: "token problem; token is malformed: token contains an invalid\
                  \ number of segments"
                path: /api/v1/product
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/product
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't update product information
                path: /api/v1/product
        "502":
          description: Ошибка при работе с api Wildberries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 502
                error: "Wildberries API Error: method not allowed"
                message: allowed methods are listed in the Allow header (POST)
                path: /api/v1/product
      security:
      - TelegramAuth: []
...
  /api/v1/product/{productId}:
    get:
      tags:
      - Product API
      summary: Получение товара
      description: Возвращает информацию о товаре пользователя
      operationId: getProduct
      parameters:
      - name: productId
        in: path
        description: Id товара
        required: true
        schema:
          type: integer
          format: int64
          example: 10000420
      responses:
        "200":
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllProductsResponse"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T20:34:31.029+00:00
                status: 400
                error: Bad Request
                path: /api/v1/product/abcd
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/product/10000420
        "403":
          description: Доступ к товару запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T20:36:51.87Z
                status: 403
                error: Access denied
                message: There is no access to this product
                path: /api/v1/product/10016160
        "404":
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T20:36:07.95Z
                status: 404
                error: No data found
                message: "Product with id: 1000000 not found"
                path: /api/v1/product/1000000
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't get product
                path: /api/v1/product/10000005
      security:
      - TelegramAuth: []
...
components:
  schemas:
    UpdateResponse:
      type: object
      description: Ответ на обновление
      properties:
        timestamp:
          type: string
          description: Дата и время обновления
          example: 2025-07-28T20:43:27.13Z
        message:
          type: string
          description: Сообщение об успехе
          example: Successfully updated the products
    AppErrorResponse:
      type: object
      description: Стандартный ответ ошибки
      properties:
        timestamp:
          type: string
          description: "Дата и время, когда произошла ошибка"
          example: 2025-07-28T11:51:39.39Z
        status:
          type: integer
          format: int32
          description: Статусный код HTTP
          example: 400
        error:
          type: string
          description: Заголовок ошибки
          example: Validation error
        message:
          type: string
          description: Описание ошибки
          example: Incorrect data was entered
        path:
          type: string
          description: "Путь запроса, где произошла ошибка"
          example: /api/v1/shipment
        details:
          type: object
          additionalProperties:
            type: string
          description: Дополнительные сведения при ошибках валидации
          example:
            quantity: The number must be greater than 0
            warehouseId: The warehouse id cannot be empty
            warehouseName: The warehouse name cannot be empty
...
    AllProductsResponse:
      type: object
      description: Ответ на получение товаров
      properties:
        products:
          type: array
          description: Список товаров
          items:
            $ref: "#/components/schemas/ProductResponse"
        size:
          type: integer
          format: int32
          description: Количество товаров в ответе
          example: 1
        hasNext:
          type: boolean
          description: Наличие товаров на следующей странице
          example: false
    ProductResponse:
      type: object
      description: Ответ с информацией о товаре
      properties:
        id:
          type: integer
          format: int64
          description: Id товара
          example: 10000420
        title:
          type: string
          description: Наименование товара
          example: Кюлоты палаццо летние
        barcode:
          type: string
          description: Баркод товара
          example: 2037989779625
        article:
          type: string
          description: Артикул товара
          example: кюлоты 400 черные
        size:
          type: string
          description: Размер товара
          example: 46
        color:
          type: string
          description: Цвет товара
          example: черный
        photoUrl:
          type: string
          description: Ссылка на первое фото товара на Wildberries
          example: https://basket-11.wbbasket.ru/vol1646/part164680/164680495/images/big/1.webp
...
  securitySchemes:
    TelegramAuth:
      type: apiKey
      description: Telegram WebApp initData
      name: Authorization
      in: header
    ApiSecretAuth:
      type: apiKey
      description: Секретный ключ API
      name: X-API-SECRET
      in: header